+++ CHANGELOG.md
@@ -13,6 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities
 
+## [1.0.1] - Upcoming
+
+### Fixed
+
+- Bugfix in `listTenants` method
+
 ## [1.0.0] - 2025.01.09
 
 ### Added